Basic .csmrc parsing
We rely here on `serde_yaml_ng`, which is a fork of `serde_yaml` that
seems at least somewhat actively maintained and used. The benefit here
is that we get easy deserialization into Rust types (which we wouldn't
get with something like `yaml_rust2` which just gives an AST). The
disadvantage is that we're depending on a library that I wish had more
community backing like the original `serde_yaml` did.

The alternative here would probably be foregoing YAML entirely and using
TOML instead, but it seems the rest of the Robotmk ecosystem uses YAML,
so it is probably best to remain consistent here.