Command: csm init
The start of shell work for the shell hook we'll need for activate and
deactivate.

As a starting point, we use clap_complete (part of clap, but a separate
crate), for shell completions. This was a bit of a side-quest, but the
main purpose is that this gives us a bounded set of shells to support,
we'll simply support the ones that clap_complete supports, for now.

Now that the infrastructure is in place, we can write the necessary
shell hooks for the shells we support.