Drop bash.exe support
On Windows we'll support PowerShell and WSL (which we should test at
some point), but let's drop bash.exe (e.g. git-bash and similar).
Supporting this properly is painful mainly because of rewriting that the
shell does internally to convert from e.g. C:\foo to /c/foo. We'd have
to do that path rewrite ourselves, from what I can tell, and it would
end up being very fragile. For now, it is likely not worth supporting
this case - most Windows users are going to be using PowerShell when
they use csm. The few who aren't, are almost certainly going to be
running it from within WSL, which should already just work.